+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1756 Fix memory exhausted during searching 
 -   \#1781 Fix search hang with SQ8H
 -   \#1818 Duplicate data generated after restart milvus server
+-   \#1827 Combine request target vectors exceed max nq
 
 ## Feature
 -   \#1603 BinaryFlat add 2 Metric: Substructure and Superstructure